requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.1
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
chromadb>=0.4.0
faiss-cpu>=1.7.4
qdrant-client>=1.7.0
pinecone-client>=3.0.0
azure-search-documents>=11.4.0
azure-core>=1.29.0
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0